Add files via upload
This repository contains the code for a project that predicts the quality of red wine. The project uses the UCI Wine Quality dataset, which contains 1599 red wine samples with 11 physicochemical properties. The project is written in Python and uses the Jupyter Notebook environment. The code is also well-documented and includes instructions on how to run it.

The project uses a variety of machine learning algorithms to predict the quality of red wine. The algorithms include random forests, support vector machines, and neural networks. The algorithms are trained on the training set and the predictions are evaluated on the test set. The best performing algorithm is then used to make predictions on new wine samples.

The project successfully predicts the quality of red wine using machine learning algorithms. The project is well-documented and includes instructions on how to run it. The project is a great resource for anyone interested in learning about machine learning or predicting the quality of red wine.